Refactor code structure
