+++ oidc-client-app-samples/ios-client-app-sample/WSO2-IS-SampleApp/utils/OAuthUtils.swift
@@ -25,119 +25,6 @@ class OAuthUtils {
     
     private init(){}
     
-    /// Retrieves user information from the server using the access token.
-    ///
-    /// - Parameters:
-    ///   - userInfoEP: User information endpoint.
-    /// - Returns: User information as a JSON object.
-    func retrieveUserInfo(userInfoURL: URL, authState: OIDAuthState?,
-                          completion: @escaping (_ userInfo : UserInfo?) -> Void) {
-        
-        // Retrieve access token from current state
-        let currentAccessToken: String? = authState?.lastTokenResponse?.accessToken
-        
-        var jsonResponse: [String: Any]?
-        
-        // Attempt to fetch fresh tokens if current tokens are expired and perform user info retrieval
-        authState?.performAction() { (accessToken, idToken, error) in
-            
-            if error != nil  {
-                print(NSLocalizedString("error.fetch.freshtokens",
-                                        comment: Constants.ErrorMessages.kErrorFetchingFreshTokens) +
-                    " \(error?.localizedDescription ?? Constants.LogTags.kError)")
-                return
-            }
-            
-            guard let accessToken = accessToken else {
-                print(NSLocalizedString("error.fetch.accesstoken",
-                                        comment: Constants.ErrorMessages.kErrorRetrievingAccessToken))
-                return
-            }
-            
-            if currentAccessToken != accessToken {
-                print(NSLocalizedString("info.accesstoken.refreshed",
-                                        comment: Constants.LogInfoMessages.kAccessTokenRefreshed) +
-                    ": (\(currentAccessToken ?? Constants.LogTags.kCurrentAccessToken) to \(accessToken))")
-            } else {
-                print(NSLocalizedString("info.accesstoken.valid",
-                                        comment: Constants.LogInfoMessages.kAccessTokenValid) + ": \(accessToken)")
-            }
-            
-            // Build user info request
-            var urlRequest = URLRequest(url: userInfoURL)
-            urlRequest.httpMethod = "GET"
-            // Request header
-            urlRequest.allHTTPHeaderFields = ["Authorization":"Bearer \(accessToken)"]
-            
-            // Retrieve user information
-            let task = URLSession.shared.dataTask(with: urlRequest) { data, response, error in
-                
-                DispatchQueue.main.async {
-                    
-                    guard error == nil else {
-                        print(NSLocalizedString("error.http.request.fail",
-                                                comment: Constants.ErrorMessages.kHTTPRequestFailed) +
-                            " \(error?.localizedDescription ?? Constants.LogTags.kError)")
-                        return
-                    }
-                    
-                    // Check for non-HTTP response
-                    guard let response = response as? HTTPURLResponse else {
-                        print(NSLocalizedString("info.nonhttp.response",
-                                                comment: Constants.LogInfoMessages.kNonHTTPResponse))
-                        return
-                    }
-                    
-                    // Check for empty response
-                    guard let data = data else {
-                        print(NSLocalizedString("info.http.response.empty",
-                                                comment: Constants.LogInfoMessages.kHTTPResponseEmpty))
-                        return
-                    }
-                    
-                    // Parse data into a JSON object
-                    do {
-                        jsonResponse = try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
-                    } catch {
-                        print(NSLocalizedString("error.json.serialization",
-                                                comment: Constants.ErrorMessages.kJSONSerializationError))
-                    }
-                    
-                    // Response with an error
-                    if response.statusCode != Constants.HTTPResponseCodes.kOk {
-                        let responseText: String? = String(data: data, encoding: String.Encoding.utf8)
-                        
-                        if response.statusCode == Constants.HTTPResponseCodes.kUnauthorised {
-                            // Possible an issue with the authorization grant
-                            let oAuthError =
-                                OIDErrorUtilities.resourceServerAuthorizationError(withCode: 0,
-                                                                                   errorResponse: jsonResponse,
-                                                                                            underlyingError: error)
-                            authState?.update(withAuthorizationError: oAuthError)
-                            print(Constants.ErrorMessages.kAuthorizationError +
-                                " (\(oAuthError)). Response: \(responseText ?? Constants.LogTags.kResponseText)")
-                        } else {
-                            print(Constants.LogTags.kHTTP +
-                                "\(response.statusCode), Response: \(responseText ?? Constants.LogTags.kResponseText)")
-                        }
-                    }
-                    
-                    if let json = jsonResponse {
-                        print(NSLocalizedString("info.information.fetch.success",
-                                                comment:Constants.LogInfoMessages.kInfoRetrievalSuccess) + ": \(json)")
-                        let userName = jsonResponse!["sub"] as? String
-                        var userInfo: UserInfo?
-                        if let un = userName {
-                            userInfo = UserInfo(userName: un)
-                        }
-                        completion(userInfo)
-                    }
-                }
-            }
-            task.resume()
-        }
-    }
-    
     /// Decodes JWT tokens.
     ///
     /// - Parameter jwt: JWT Token string.
